Refactor YouTube audio downloader to call yt_dlp
This commit refactors the Youtube-mp3-downloader to call `yt_dlp`
directly from the program and store the data in memory. Previously, the
program relied on the `yt-dlp` library, but now it uses the `yt_dlp`
package for improved compatibility and control.

One of the challenges encountered was that `yt_dlp` cannot convert WebM
to MP3 when outputting to stdout. To address this limitation, we have
integrated `pydub` into the process. After downloading the audio in
WebM format, `pydub` is used to convert it to MP3 before sending it to
the user.

By adopting this approach, we prevent multiple users from downloading
the same YouTube video simultaneously, which could potentially lead to
race conditions. This is achieved by avoiding conflicts in file names
since each user receives a unique MP3 file.